# Conflicts:
#	app/src/main/java/com/edifica/fragments/clients/FragmentClientBudgets.kt
#	app/src/main/res/layouts/activities/clients/layout/activity_client_chat.xml

+++ app/src/main/java/com/edifica/fragments/clients/FragmentClientBudgets.kt
@@ -1,6 +1,5 @@
 package com.edifica.fragments.clients
 
-import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import androidx.fragment.app.Fragment
@@ -11,11 +10,10 @@ import android.widget.TextView
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.edifica.R
-import com.edifica.activities.clients.ActivityClientChat
-import com.edifica.activities.clients.ActivityClientFormulary
 import com.edifica.adapters.TransactionsClientAdapter
 import com.edifica.interfaces.TransactionListener
 import com.edifica.models.*
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.DocumentReference
 import com.google.firebase.firestore.FirebaseFirestore
@@ -104,11 +102,25 @@ class FragmentClientBudgets : Fragment(), TransactionListener {
     }
 
     override fun cancelOnItemClick(transaction: Transactions, position: Int) {
-        db.collection("transaction").document(transaction.id).delete()
 
-        transactions.remove(transaction)
+        val dialogBuilder = MaterialAlertDialogBuilder(context, R.style.MyMaterialAlertDialog)
 
-        loadAdapter()
+        dialogBuilder.setMessage(R.string.fragment_client_profile_chat_question_delete_alert_dialog)
+            .setPositiveButton(R.string.fragment_client_profile_chat_affirmative_delete_alert_dialog) { _, _ ->
+                db.collection("transaction").document(transaction.id).delete()
+
+                transactions.remove(transaction)
+
+                loadAdapter()
+            }
+            .setNegativeButton(R.string.fragment_client_profile_chat_negative_delete_alert_dialog) { dialog, _ ->
+                dialog.cancel()
+            }
+
+        val alert = dialogBuilder.create()
+
+        alert.setTitle(R.string.fragment_client_profile_chat_action_delete_alert_dialog)
+        alert.show()
     }
 
     override fun chatOnItemClick(transaction: Transactions, position: Int) {
